libcluu: store cells as u32 codepoints with UTF-8 decoder
Widen Event::Print and per-cell storage from u8 (raw CP437 byte) to u32
(Unicode codepoint) so TUI apps can emit UTF-8 multi-byte sequences and
have them survive the cluuterm → compositor pipeline.

- ansi parser gains an Utf8 state that decodes 2/3/4-byte leads and
  emits a single Event::Print(cp); malformed tails emit U+FFFD.
- cluuterm cells/scrollback HistoryRow.chars move to Vec<u32>; pack_cell
  takes cp: u32 to match compositor's compose::pack_cell.
- compositor unicode_to_cp437 extended: ├ ┤ ┴ ┬ ┼, half blocks ▀▄▌▐,
  shade blocks ░▒▓, on top of the existing box corners.
- console renderer updated for the new u32 codepoint flow.

+++ userspace/cluuterm/src/render.rs
@@ -108,7 +108,7 @@ pub fn argb_to_palette_idx(argb: u32) -> u8 {
 /// Pack `(codepoint:21, fg:8, bg:8, attrs:3)` into a single u64 cell word.
 /// Matches the compositor's `compose::pack_cell` format exactly.
 #[inline]
-fn pack_cell(cp: u8, fg: u8, bg: u8, attrs: u8) -> u64 {
+fn pack_cell(cp: u32, fg: u8, bg: u8, attrs: u8) -> u64 {
     (cp as u64 & 0x1F_FFFF)
         | ((fg as u64 & 0xFF) << 21)
         | ((bg as u64 & 0xFF) << 29)
